Reorganized batch operations into batch_operations/ subdirectories. Added tests.
