--- batch-operations/java/pgjdbc/build.gradle
@@ -1,27 +1,0 @@
-plugins {
-    id 'java'
-    id 'application'
-}
-
-group = 'com.example.dsql'
-version = '1.0.0'
-
-java {
-    sourceCompatibility = JavaVersion.VERSION_17
-    targetCompatibility = JavaVersion.VERSION_17
-}
-
-repositories {
-    mavenCentral()
-}
-
-dependencies {
-    implementation 'org.postgresql:postgresql:42.7.10'
-    implementation 'com.zaxxer:HikariCP:7.0.2'
-    implementation 'software.amazon.dsql:aurora-dsql-jdbc-connector:1.4.0'
-    implementation 'software.amazon.awssdk:dsql:2.42.14'
-}
-
-application {
-    mainClass = 'com.example.dsql.Main'
-}
